Fumbling about with the managed-land component list and destruction workflow
Decided the original idea went too much against the engine instead of with it. And destroys are always called at the end of a frame anyway so there was no need to defer it further.

Added some more tests that deal with actor destruction

Changed some tests because destruction workflow changed.