Putting a bandaid on Unitys flesh wound
It's really gunky but there you go.

We will now re-trigger all the bound axis every frame for as long as there is any axis with a value != 0.

To come by the crappyness of having to use a tick function here, it will only tick when there are any non-0 inputs available.