Ironing out little quirks
Made GameComponent Awake virtual - I would avoid using it but especially on systems where broilerplate is tacked on as a refactoring operation this may become a necessary crutch.

Adding PlayerController instance to the list of players before instantiating PlayerPawn.
This ensures PlayerController is accessible for pawn and components when they spawn.