Added Timer tool
This is different from an interpolator in that it actually ticks.

This is different from the Interpolator.Anim8 in that it doesn't use coroutines but the games tick loop.

Additionally, this allows multiple listeners to the different timer events.

Naturally, this means that this timer cannot survive scene swaps as it is bound to the world.

But it is a quick and easy way to do anything on a timer with any interval.